requirements.txt
# Chess Agent - Requirements
# For Apple Silicon Mac: pip install -r requirements.txt

torch>=2.1.0
numpy>=1.24.0
python-chess>=1.999
pyyaml>=6.0
tqdm>=4.65.0
berserk>=0.13.0
tensorboard>=2.14.0
